Remove some references to longs

+++ ssh/dsskey.py
@@ -79,7 +79,6 @@ class DSSKey (PKey):
         h = h * 37 + hash(self.q)
         h = h * 37 + hash(self.g)
         h = h * 37 + hash(self.y)
-        # h might be a long by now...
         return hash(h)
 
     def get_name(self):
@@ -93,7 +92,7 @@ class DSSKey (PKey):
 
     def sign_ssh_data(self, rng, data):
         digest = SHA.new(data).digest()
-        dss = DSA.construct((long(self.y), long(self.g), long(self.p), long(self.q), long(self.x)))
+        dss = DSA.construct((self.y, self.g, self.p, self.q, self.x))
         # generate a suitable k
         qsize = len(util.deflate_long(self.q, 0))
         while True:
@@ -128,7 +127,7 @@ class DSSKey (PKey):
         sigS = util.inflate_long(sig[20:], 1)
         sigM = util.inflate_long(SHA.new(data).digest(), 1)
 
-        dss = DSA.construct((long(self.y), long(self.g), long(self.p), long(self.q)))
+        dss = DSA.construct((self.y, self.g, self.p, self.q))
         return dss.verify(sigM, (sigR, sigS))
 
     def _encode_key(self):
